Support limiting verbose logs by logger name
When creating a new verbose logger, not only check if the given
verbosity level is less than or equal to the configured verbosity
setting, but also check to see if the given logger name is included in
the list of loggers to limit by. If the list of loggers to limit by is
empty, or if the current logger is included, then an enabled logger is
returned. Otherwise, a disabled logger is returned. As before, if the
given verbosity level is greater than the configured verbosity setting,
a disabled logger is returned.

Also moved verbosity setting to be at the package level so it can be set
after loggers have already been created. This allows for package loggers
to be created as package variables.

Also added human-readable timestamp to log message. `ts` is now
human-readable, and `epoch` is, well... what it sounds like.